+++ src/Schema/AST/ASTBuilder.php
@@ -21,7 +21,7 @@ use Nuwave\Lighthouse\Schema\RootType;
 use Nuwave\Lighthouse\Schema\Source\SchemaSourceProvider;
 use Nuwave\Lighthouse\Support\Contracts\ArgManipulator;
 use Nuwave\Lighthouse\Support\Contracts\FieldManipulator;
-use Nuwave\Lighthouse\Support\Contracts\InputManipulator;
+use Nuwave\Lighthouse\Support\Contracts\InputFieldManipulator;
 use Nuwave\Lighthouse\Support\Contracts\TypeExtensionManipulator;
 use Nuwave\Lighthouse\Support\Contracts\TypeManipulator;
 
@@ -74,7 +74,7 @@ class ASTBuilder
         $this->applyTypeExtensionManipulators();
         $this->applyFieldManipulators();
         $this->applyArgManipulators();
-        $this->applyInputManipulators();
+        $this->applyInputFieldManipulators();
 
         // Listeners may manipulate the DocumentAST that is passed by reference
         // into the ManipulateAST event. This can be useful for extensions
@@ -227,15 +227,15 @@ class ASTBuilder
     }    
     
     /** Apply directives on input fields that can manipulate the AST. */
-    protected function applyInputManipulators(): void
+    protected function applyInputFieldManipulators(): void
     {
         foreach ($this->documentAST->types as $typeDefinition) {
             if ($typeDefinition instanceof InputObjectTypeDefinitionNode) {
                 foreach ($typeDefinition->fields as $fieldDefinition) {
                     foreach (
-                        $this->directiveLocator->associatedOfType($fieldDefinition, InputManipulator::class) as $inputManipulator
+                        $this->directiveLocator->associatedOfType($fieldDefinition, InputFieldManipulator::class) as $inputFieldManipulator
                     ) {
-                        $inputManipulator->manipulateInputDefinition(
+                        $inputFieldManipulator->manipulateInputFieldDefinition(
                             $this->documentAST,
                             $fieldDefinition,
                             $typeDefinition,